requirements.txt
# General
telethon
pyrogram==1.1.13
aiogram

# DBs
redis
aioredis # Redis memery storage fom aiogram
pymongo
motor
sqlalchemy==1.3.20 # For old modules
dnspython # needed to connect to cloud MongoDB instances


# Optional deps to make bot faster
aiohttp[speedups]
cryptg
cryptography

# Other
jikanpy
envparse
hypercorn
aiocron
apscheduler
requests
python-rapidjson
PyYAML>5.0
coloredlogs
loguru
babel
captcha
async-timeout
regex
bs4
lxml
spamwatch
httpx[http2]
wikipedia
coffeehouse
PyDictionary==2.0.1
google-trans-new==1.1.9
hachoir
telegraph
faker
gtts
geopy
tswift
lyricsgenius
sentry_sdk
lxml
html5lib
feedparser
tldextract
dateparser

#Pyrogram
tgCrypto>=1.2.2

#Old DB Processes
psycopg2-binary==2.8.6

#YTDL
youtube_dl
youtube_search_python

#Search
bs4
html2text
bing_image_downloader 
search_engine_parser
pornhub_api
youtube_search

#Memes
selenium
zalgo_text 
cowpy 
fontTools
nltk
emoji

#Deezer
wget
asyncio

#Markup
Markdown>=3.3.4

#profanity (Credits to Julia)
better_profanity
textblob 
nudepy

#ImageEditor
glitch_this
NumPy
opencv-python-headless
Pillow

#Bassboost
pydub

#Lel
cloudmersive_virus_api_client==3.0.1
chromedriver==2.24.1

#Updator
gitpython==3.1.11
heroku3==4.2.3

pykeyboard 
countryinfo
flag
Python_ARQ 
googletrans==4.0.0-rc1
pyromod
